+++ django/contrib/admindocs/views.py
@@ -30,7 +30,7 @@ from django.utils.inspect import (
 from django.utils.translation import gettext as _
 from django.views.generic import TemplateView
 
-from .utils import _is_callback, get_view_name
+from .utils import get_view_name
 
 # Exclude methods starting with these strings from documentation
 MODEL_METHODS_EXCLUDE = ("_", "add_", "delete", "save", "set_")
@@ -166,7 +166,8 @@ class ViewDetailView(BaseAdminDocsView):
 
     @staticmethod
     def _get_view_func(view):
-        if _is_callback(view):
+        urlconf = get_urlconf()
+        if get_resolver(urlconf)._is_callback(view):
             mod, func = get_mod_func(view)
             try:
                 # Separate the module and function, e.g.
